refactor: userInput 클래스 분리
